Projeto Lima Consulting/requirements.txt
# Bibliotecas utilizadas no projeto:
selenium
pyspark
Scrapy
google-cloud-bigquery
